+++ lib/domain/model_type_id.dart
@@ -101,5 +101,5 @@ class ModelTypeId {
   static const mediaType = 79;
   static const hiveBackground = 80;
   static const storageFile = 81;
-  static const typedInChat = 82;
+  static const draftMessage = 82;
 }