+++ test/e2e/parameters/keys.dart
@@ -41,6 +41,7 @@ enum WidgetKey {
   Delete,
   DeleteAccountButton,
   DeleteAvatar,
+  DraftMessage,
   EmailsExpandable,
   ForwardButton,
   ForwardField,